Made the NeuronServer work well enough to create a list of open items. Also fixed some bugs introduced in the label cacheing.

+++ BrainSimulator/NeuronArray.cs
@@ -5,6 +5,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
+using System.Xml.Serialization;
 
 namespace BrainSimulator
 {
@@ -17,7 +18,6 @@ namespace BrainSimulator
         public bool EngineIsPaused = false;
         public int arraySize;
         public int rows;
-
 
         public int lastFireCount = 0;
         internal List<ModuleView> modules = new List<ModuleView>();
@@ -45,6 +45,14 @@ namespace BrainSimulator
                 labelCache[nID] = label;
             }
         }
+        public void RemoveLabelFromCache(int nID)
+        {
+            try
+            {
+                labelCache.Remove(nID);
+            }
+            catch { };
+        }
         public string GetLabelFromCache(int nID)
         {
             if (labelCache.ContainsKey(nID))
@@ -70,24 +78,26 @@ namespace BrainSimulator
         public int RefractoryDelay
         { get => refractoryDelay; set { refractoryDelay = value; SetRefractoryDelay(refractoryDelay); } }
 
-        public void Initialize(int count, int inRows)
+        public void Initialize(int count, int inRows,bool clipBoard = false)
         {
             rows = inRows;
             arraySize = count;
             ClearLabelCache();
-            if (!MainWindow.useServers)
+            if (!MainWindow.useServers || clipBoard)
                 base.Initialize(count);
+            else
+            {
+                NeuronClient.InitServers(0, count);
+            }
         }
-
-
 
         public NeuronArray()
         {
         }
 
-        public new Neuron GetNeuron(int id)
+        public new Neuron GetNeuron(int id,bool fromClipboard = false)
         {
-            Neuron n = GetCompleteNeuron(id);
+            Neuron n = GetCompleteNeuron(id,fromClipboard);
             return n;
         }
         public Neuron GetNeuron(string label)
@@ -97,19 +107,22 @@ namespace BrainSimulator
                 int nID = labelCache.FirstOrDefault(x => x.Value == label).Key;
                 return GetNeuron(nID);
             }
-            //for (int i = 0; i < arraySize; i++)
-            //{
-            //    Neuron n = GetNeuron(i);
-            //    if (n.label == label) return n;
-            //}
             return null;
         }
 
 
         public void GetCounts(out long synapseCount, out int useCount)
         {
-            synapseCount = GetTotalSynapses(); ;
-            useCount = GetTotalNeuronsInUse();
+            if (MainWindow.useServers)
+            {
+                useCount = NeuronClient.serverList.Sum(x => x.neuronsInUse);
+                synapseCount = NeuronClient.serverList.Sum(x => x.totalSynapses);
+            }
+            else
+            {
+                synapseCount = GetTotalSynapses();
+                useCount = GetTotalNeuronsInUse();
+            }
         }
 
         public new void Fire()
@@ -133,14 +146,14 @@ namespace BrainSimulator
         }
         public void AddSynapse(int src, int dest, float weight, Synapse.modelType model, bool noBackPtr)
         {
-            if (MainWindow.useServers)
+            if (MainWindow.useServers && this == MainWindow.theNeuronArray)
                 NeuronClient.AddSynapse(src, dest, weight, model, noBackPtr);
             else
                 base.AddSynapse(src, dest, weight, (int)model, noBackPtr);
         }
         new public void DeleteSynapse(int src, int dest)
         {
-            if (MainWindow.useServers)
+            if (MainWindow.useServers && this == MainWindow.theNeuronArray)
                 NeuronClient.DeleteSynapse(src, dest);
             else
                 base.DeleteSynapse(src, dest);
